chore: mv test to internal
