+++ __tests__/02.patch-state.spec.ts
@@ -14,15 +14,7 @@
 /* eslint-disable @typescript-eslint/no-extraneous-class, @typescript-eslint/no-unsafe-member-access */
 import path from 'path';
 import * as borsh from 'borsh';
-import {NEAR} from 'near-units';
 import {Runner} from '../src';
-import {RecordBuilder} from '../src/record';
-
-async function sleep(ms: number): Promise<boolean> {
-  return new Promise(resolve => {
-    setTimeout(() => resolve(true), ms); // eslint-disable-line @typescript-eslint/no-confusing-void-expression
-  });
-}
 
 describe('view state & patch state', () => {
   if (Runner.networkIsSandbox()) {
@@ -109,38 +101,6 @@ describe('view state & patch state', () => {
           account_id: 'alice.near',
         });
         expect(result).toBe('hello world');
-      });
-    });
-
-    test.concurrent('Patch Account', async () => {
-      await runner.run(async ({root, ali, contract}) => {
-        // Return;
-        const bob = root.getFullAccount('bob');
-        const public_key = await bob.setKey();
-        const {code_hash} = await contract.accountView();
-        const BOB_BALANCE = NEAR.parse('100 N');
-        const rb = RecordBuilder.fromAccount(bob.accountId)
-          .account({
-            amount: BOB_BALANCE.toString(),
-            code_hash,
-          }).accessKey(
-            public_key,
-            {
-              nonce: 0,
-              permission: 'FullAccess',
-            },
-          )
-          .contract(await contract.viewCode());
-        await bob.sandbox_patch_state(rb);
-        await sleep(1000);
-        await bob.sandbox_patch_state(rb);
-        const balance = await bob.availableBalance();
-        expect(balance).toStrictEqual(BOB_BALANCE);
-        await ali.call(bob, 'set_status', {message: 'hello'});
-        const result = await bob.view('get_status', {
-          account_id: ali.accountId,
-        });
-        expect(result).toBe('hello');
       });
     });
   } else {
